+++ tests/testthat/_snaps/rix-julia/julia_default.nix
@@ -1,6 +1,6 @@
 
 let
-  pkgs = import (fetchTarball "https://github.com/rstats-on-nix/nixpkgs/archive/2025-03-10.tar.gz") {};
+  pkgs = import (fetchTarball "https://github.com/rstats-on-nix/nixpkgs/archive/2025-09-04.tar.gz") {};
  
   rpkgs = builtins.attrValues {
     inherit (pkgs.rPackages)