+++ tests/testthat/test-rix-julia.R
@@ -22,7 +22,7 @@ testthat::test_that("rix() with Julia packages", {
 
   save_default_nix_test <- function(ide, path_default_nix) {
     rix(
-      date = "2025-03-10",
+      date = "2025-09-04",
       r_pkgs = c("dplyr", "janitor", "reticulate"),
       tex_pkgs = c("amsmath"),
       jl_conf = list(
@@ -39,11 +39,50 @@ testthat::test_that("rix() with Julia packages", {
     file.path(path_default_nix, "default.nix")
   }
 
-  suppressWarnings(
-    testthat::expect_snapshot_file(
-      path = save_default_nix_test(ide = "none", path_default_nix),
-      name = "julia_default.nix",
-    )
+  testthat::expect_snapshot_file(
+    path = save_default_nix_test(ide = "none", path_default_nix),
+    name = "julia_default.nix",
+  )
+})
+
+testthat::test_that("rix() with Julia packages, older date", {
+  os_type <- Sys.info()["sysname"]
+  skip_if_not(nix_shell_available())
+  skip_if(os_type == "Darwin" || os_type == "Windows")
+
+  tmpdir <- tempdir()
+
+  path_default_nix <- paste0(
+    tmpdir,
+    paste0(sample(letters, 5), collapse = "")
+  )
+  dir.create(path_default_nix)
+  path_default_nix <- normalizePath(path_default_nix)
+  on.exit(
+    unlink(path_default_nix, recursive = TRUE, force = TRUE),
+    add = TRUE
+  )
+  on.exit(
+    unlink(tmpdir, recursive = TRUE, force = TRUE),
+    add = TRUE
+  )
+
+  testthat::expect_warning(
+    rix(
+      date = "2025-04-04",
+      r_pkgs = c("dplyr", "janitor", "reticulate"),
+      tex_pkgs = c("amsmath"),
+      jl_conf = list(
+        jl_version = "1.10",
+        jl_pkgs = c("TidierData", "RDatasets")
+      ),
+      ide = ide,
+      project_path = path_default_nix,
+      overwrite = TRUE,
+      message_type = "quiet",
+      shell_hook = NULL
+    ),
+    "Julia support is only guaranteed from 2025-09-04 onward.",
   )
 })
 
@@ -72,7 +111,7 @@ testthat::test_that("rix() with lts julia version", {
 
   save_default_nix_test <- function(ide, path_default_nix) {
     rix(
-      date = "2025-03-10",
+      date = "2025-09-04",
       r_pkgs = c("dplyr", "janitor", "reticulate"),
       tex_pkgs = c("amsmath"),
       jl_conf = list(
@@ -89,11 +128,9 @@ testthat::test_that("rix() with lts julia version", {
     file.path(path_default_nix, "default.nix")
   }
 
-  suppressWarnings(
-    testthat::expect_snapshot_file(
-      path = save_default_nix_test(ide = "none", path_default_nix),
-      name = "julia_default_lts_version.nix",
-    )
+  testthat::expect_snapshot_file(
+    path = save_default_nix_test(ide = "none", path_default_nix),
+    name = "julia_default_lts_version.nix",
   )
 })
 
@@ -122,7 +159,7 @@ testthat::test_that("rix() with empty julia version", {
 
   save_default_nix_test <- function(ide, path_default_nix) {
     rix(
-      date = "2025-03-10",
+      date = "2025-09-04",
       r_pkgs = c("dplyr", "janitor", "reticulate"),
       tex_pkgs = c("amsmath"),
       jl_conf = list(
@@ -138,10 +175,8 @@ testthat::test_that("rix() with empty julia version", {
     file.path(path_default_nix, "default.nix")
   }
 
-  suppressWarnings(
-    testthat::expect_snapshot_file(
-      path = save_default_nix_test(ide = "none", path_default_nix),
-      name = "julia_default_empty_version.nix",
-    )
+  testthat::expect_snapshot_file(
+    path = save_default_nix_test(ide = "none", path_default_nix),
+    name = "julia_default_empty_version.nix",
   )
 })